Migrate MO Surfaces to MolecularSurfaces Layer

+++ src/Data/GeminalOrbitals.h
@@ -54,7 +54,7 @@ namespace Data {
 
          double geminalOrbitalEnergy(unsigned i) const 
          { 
-            return (i < m_geminalEnergies.length()) ? m_geminalEnergies[i] : 0.0;
+            return ((int)i < m_geminalEnergies.length()) ? m_geminalEnergies[i] : 0.0;
          }
 
          void boundingBox(qglviewer::Vec& min, qglviewer::Vec& max) const
